Switch to a modular layout.
